display created, updated at as UTC
The original data type for these fields in oracle was DATE, and I was storing SYSDATE in that field when an inventory was created. When querying this field, it would return the date in the local timezone. DATE doesn't support AT TIME ZONE, so I had to change the type to TIMESTAMP and modify the query to get the data as UTC